pyproject.toml
[project]
name = "svi-to-poi"
version = "0.1.0"
description = "HCMC Spatial-Semantic POI Reconstruction from Street View Images"
readme = "CLAUDE.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Project Team"}
]
keywords = ["poi", "street-view", "geospatial", "vlm", "ho-chi-minh-city"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    # Core Data
    "pandas>=2.0.0",
    "numpy>=1.24.0",

    # Geospatial
    "geopandas>=0.14.0",
    "shapely>=2.0.0",
    "pyproj>=3.6.0",
    "folium>=0.15.0",
    "mercantile>=1.2.0",
    "vt2geojson>=0.2.0",

    # Image Processing
    "opencv-python>=4.8.0",
    "Pillow>=10.0.0",

    # ML/DL
    "torch>=2.0.0",
    "transformers>=4.35.0",

    # Network/API
    "requests>=2.31.0",
    "httpx>=0.25.0",
    "tenacity>=8.2.0",

    # CLI/Utils
    "click>=8.1.0",
    "tqdm>=4.66.0",
    "python-dotenv>=1.0.0",

    # VLM Clients
    "openai>=1.0.0",

    # Visualization
    "plotly>=5.18.0",

    # Type Checking
    "pydantic>=2.5.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.21.0",
    "pytest-mock>=3.12.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.uv]
dev-dependencies = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.21.0",
    "pytest-mock>=3.12.0",
    "ruff>=0.1.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --strict-markers --cov=src --cov-report=term-missing"

[tool.coverage.run]
source = ["src"]
omit = ["*/tests/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
]
